refactor: rearrange mods, fixed typos and added serialize impl
