+++ fixtures/r4/resources/location.json
@@ -1,0 +1,85 @@
+{
+  "resourceType": "Location",
+  "id": "location-1",
+  "identifier": [
+    {
+      "system": "http://example.com/locations",
+      "value": "loc-001"
+    }
+  ],
+  "status": "active",
+  "operationalStatus": {
+    "system": "http://terminology.hl7.org/CodeSystem/v2-0116",
+    "code": "C",
+    "display": "Closed"
+  },
+  "name": "South Wing, second floor",
+  "alias": [
+    "SW-2"
+  ],
+  "description": "Second floor of the south wing",
+  "mode": "instance",
+  "type": [
+    {
+      "coding": [
+        {
+          "system": "http://terminology.hl7.org/CodeSystem/v3-RoleCode",
+          "code": "HOSP",
+          "display": "Hospital"
+        }
+      ],
+      "text": "Hospital"
+    }
+  ],
+  "telecom": [
+    {
+      "system": "phone",
+      "value": "2328"
+    }
+  ],
+  "address": {
+    "use": "work",
+    "line": [
+      "South Wing, floor 2"
+    ],
+    "city": "PleasantVille",
+    "state": "Vic",
+    "postalCode": "3999",
+    "country": "Australia"
+  },
+  "physicalType": {
+    "coding": [
+      {
+        "system": "http://terminology.hl7.org/CodeSystem/location-physical-type",
+        "code": "wi",
+        "display": "Wing"
+      }
+    ]
+  },
+  "position": {
+    "longitude": -83.69,
+    "latitude": 42.25,
+    "altitude": 0.0
+  },
+  "hoursOfOperation": [
+    {
+      "daysOfWeek": [
+        "mon",
+        "tue",
+        "wed"
+      ],
+      "allDay": false
+    }
+  ],
+  "availabilityExceptions": "Closed on public holidays",
+  "managingOrganization": {
+    "reference": "Organization/1",
+    "display": "Burgers University Medical Center"
+  },
+  "partOf": {
+    "reference": "Location/parent"
+  },
+  "endpoint": {
+    "reference": "Endpoint/1"
+  }
+}
