+++ fixtures/r4/resources/organization.json
@@ -1,0 +1,101 @@
+{
+  "resourceType": "Organization",
+  "id": "org-1",
+  "identifier": [
+    {
+      "use": "official",
+      "system": "http://example.org/orgs",
+      "value": "ORG-001"
+    }
+  ],
+  "active": true,
+  "type": [
+    {
+      "coding": [
+        {
+          "system": "http://terminology.hl7.org/CodeSystem/organization-type",
+          "code": "prov",
+          "display": "Healthcare Provider"
+        }
+      ],
+      "text": "Healthcare Provider"
+    }
+  ],
+  "name": "Burgers University Medical Center",
+  "alias": [
+    "BUMC",
+    "Burgers UMC"
+  ],
+  "telecom": [
+    {
+      "system": "phone",
+      "value": "+1-555-123-4567",
+      "use": "work"
+    },
+    {
+      "system": "email",
+      "value": "info@bumc.example.org",
+      "use": "work"
+    }
+  ],
+  "address": [
+    {
+      "use": "work",
+      "line": [
+        "123 Medical Way"
+      ],
+      "city": "PleasantVille",
+      "state": "Vic",
+      "postalCode": "3999",
+      "country": "Australia"
+    }
+  ],
+  "partOf": {
+    "reference": "Organization/parent-org",
+    "display": "Burgers Health Network"
+  },
+  "contact": [
+    {
+      "purpose": {
+        "coding": [
+          {
+            "system": "http://terminology.hl7.org/CodeSystem/contactentity-type",
+            "code": "ADMIN",
+            "display": "Administrative"
+          }
+        ],
+        "text": "Administrative"
+      },
+      "name": {
+        "family": "Smith",
+        "given": [
+          "Jane"
+        ]
+      },
+      "telecom": [
+        {
+          "system": "phone",
+          "value": "+1-555-987-6543",
+          "use": "work"
+        }
+      ],
+      "address": {
+        "line": [
+          "123 Medical Way"
+        ],
+        "city": "PleasantVille",
+        "state": "Vic",
+        "postalCode": "3999",
+        "country": "Australia"
+      }
+    }
+  ],
+  "endpoint": [
+    {
+      "reference": "Endpoint/endpoint-1"
+    },
+    {
+      "reference": "Endpoint/endpoint-2"
+    }
+  ]
+}
