+++ src/main.rs
@@ -1,8 +1,13 @@
-use libfhir::resources::LocationPosition;
+use libfhir::r4::resources::OrganizationBuilder;
 
 mod fhir;
 
 fn main() {
-    let position = LocationPosition::default();
-    println!("{:?}", position)
+    let org = OrganizationBuilder::new("org-1")
+        .active(true)
+        .add_alias("some alias")
+        .name("some-name")
+        .build();
+
+    println!("{:?}", org.to_json_string())
 }
