+++ src/r4/elements/complex_types/money.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::{
+use crate::r4::{
     elements::Element,
     resources::{self, ResourceType},
 };
