+++ src/r4/elements/complex_types/range.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::{
+use crate::r4::{
     elements::{Element, Quantity},
     resources::{self, ResourceType},
 };
@@ -63,7 +63,7 @@ impl RangeBuilder {
 
 #[cfg(test)]
 mod test {
-    use crate::elements::QuantityBuilder;
+    use crate::r4::elements::QuantityBuilder;
 
     use super::*;
