+++ src/r4/elements/element.rs
@@ -1,11 +1,14 @@
 use serde::{Deserialize, Serialize};
 
-use crate::resources::ResourceType;
+use crate::r4::resources::ResourceType;
 
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize, Default)]
 #[serde(rename_all(serialize = "snake_case", deserialize = "camelCase"))]
 pub struct Element {
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub id: Option<String>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub extention: Option<Vec<String>>, // to be resolved later
 }
 
@@ -13,6 +16,8 @@ pub struct Element {
 pub struct BackboneElement {
     #[serde(flatten)]
     pub element: Element,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub modifier_extensions: Option<Vec<String>>, // to be resolved later
 }
