+++ src/r4/resources/resource.rs
@@ -4,8 +4,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Default, Serialize, Deserialize, PartialEq, Clone)]
 pub struct Resource {
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub id: Option<String>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub meta: Option<String>, // to be resoloved later
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub implicit_rules: Option<String>,
 }
 
@@ -13,16 +18,22 @@ pub struct Resource {
 pub struct DomainResource {
     #[serde(flatten)]
     pub resource: Resource,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub text: Option<String>,
+
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub contained: Option<Vec<String>>, // To be resolved later
-    pub exnetions: Option<Vec<String>>, // to be resolved later
+
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub extensions: Option<Vec<String>>, // to be resolved later
 }
 
 pub trait ResourceType {
     const TYPE: &'static str;
 
-    fn get_resource_type(&self) -> &'static str {
-        Self::TYPE
+    fn get_resource_type() -> String {
+        Self::TYPE.to_string()
     }
 }
